Fix surface reconstruction when trace length is less than 5
Currently I don't get a surface reconstruction whenever this warning is
shown:
"Warning: Too few points in trace (length < 5)"
This is because the exceptional return value as the wrong signature
"return [np.nan] * 5, intensity, intensity"
Which leads to an index out of bounds in _fance_edge_fit which is not
handled properly.

This commit replaces the exceptional return value with an ValueError which
is properly handled by _fit_edge_fancy.